+++ Galary/MyGalary/GalaryRootTableViewController.m
@@ -19,14 +19,14 @@ typedef void(^CustomPickerHandler)(NSUInteger index);
     BOOL mIncrementalCount;
     NSArray * mCustomPickers;
     CustomPickerHandler mCustomPickerHandler;
-    int mMaxCount;
+    NSUInteger mMaxCount;
 }
 @property (nonatomic, strong) NSArray *sectionFetchResults;
 @end
 
 @implementation GalaryRootTableViewController
 
-- (instancetype) initWithIncrementalCount : (BOOL) incrementalCount withPickComplete : (void (^)(NSArray<PHAsset*>* assets)) pickComplete withCustomPicker : (NSArray<UIImage*>*) customPickers withCustomPickerHandler : (void (^)(NSUInteger index)) customPickerHandler maxCount : (int) maxCount
+- (instancetype) initWithIncrementalCount : (BOOL) incrementalCount withPickComplete : (void (^)(NSArray<PHAsset*>* assets)) pickComplete withCustomPicker : (NSArray<UIImage*>*) customPickers withCustomPickerHandler : (void (^)(NSUInteger index)) customPickerHandler maxCount : (NSUInteger) maxCount
 {
     self = [super init];
     if(self){
